image_steganography/requirements.txt
Django==4.2.7
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.0
